fine tunning2
